Basic weapon firing and basic reload functionality added.
Main issue: I still haven't mastered the Input System, therefore I'm unsure how to make it behave as the old Input.
This means that I am perpetually stuck on semi auto fire, and there's an annoying bug where the shot gets "saved" when the player is in "idle" state and can't fire, and then the Shoot method is called as soon as the player enters the aim state. I'm sure this is connected to the events being run in the background.