InputAction and Inputs refactoring started
Now that I have achieved very basic functionality that I was aiming for, I'm going to refactor input and tailor it to the project's specs:
Goals for the refactor:
Better scalability and readability.
Expanded functionality.
Improved responsiveness.
Improved gameplay potential.
This will allow me to finally implement automatic and select weapon fire. This is A MUST, because without it I can't implement bullet spread and accuracy mechanics I have planned from the start.