Expanded weapon functionality. Start on the fire spread functionality (also called blooming)
I finally did it! My weapons can now fire in full auto or in Semi Auto. The slight issue is that I need two different methods for firing, even though they rely  on the same method  - FireAction.
This was caused due to me being uncomfortable with the Input System, and I didn't pay attention to the Control Type in the editor. It's all fixed now, and I can fire away.
BUG: The final stage, where the player is supposed to play one shot sounds when out of ammo is broken now.
Temporary solution is going to be implemented in the next commit.