Started adding enemy detector
I found an interesting problem. If I place a trigger collider on an enemy child object, the player aim gets disrupted if I am too close to the enemy.  It think the problem is on the ThirdPersonShooterController. I have a hunch what's causing it (raycast that the aim function has as a part of it). If I can't solve it, I'll do it with another raycast from the enemy detector game object. IE, I'll solve it programatically.