Made a number of tweaks and improvements
Made a HordeSpawner script that randomly spawns a horde of randomly genrated zombies - this means that I no longer have to do it by hand, instead all I need to do is decide how many zombies to spawn and .

Added a vehicle control script (not mine - I found the code online - I was only testing a potential feature. Vehicles are not expected to be implemented before Pre-Alpha 0.3 AT LEAST).