Smooth shoulder switching added
The functionality works as I intended it to from the start.
Known bugs: It does not work as intended if the shoulderSwitch time is not exactly 1.0f. I'll need to investigate later.
Right now, the player cannot interrupt the switch during the coroutine firing off. I'll need to refactor and try to implement this functionaltiy later.

Note: Overall, I'm peasantly surprised that I managed to make the game feel more like a modern TPS in the vein of Ghost Recon Wildlands/Breakpoint.